Gates and policies

+++ app/Http/Controllers/QuestionController.php
@@ -76,7 +76,11 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return View('questions.edit',compact('question'));
+		if(\Gate::denies('update-question',$question)) {
+			abort(403,'Access denied!');
+		}
+		$this->authorize('update',$question);
+		return View('questions.edit',compact('question'));
     }
 
     /**
@@ -89,6 +93,7 @@ class QuestionController extends Controller
     public function update(Request $request, Question $question)
     {
         //
+		$this->authorize('update',$question);
 		$question->update([
 			'name' => $request->title,
 			'body' => $request->body
@@ -104,6 +109,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+		if(\Gate::denies('delete-question',$question)) {
+			abort(403,'Access denied!');
+		}		
+		$this->authorize('delete',$question);
         $question->delete();
 		return redirect('/question')->with('success','Your question has beed deleted');
     }
